Move User to Models folder
